requirements.txt
tinkoff-investments
aiogram
pandas>=2.0.0
pandas-ta>=0.3.14b0